+++ Assets/Scripts/Units/Abilities/AttackEffects/Damage/DamagePerShield.cs
@@ -3,7 +3,7 @@
 [CreateAssetMenu(fileName = "Per Shield Mod", menuName = "Ability/Attack/Damage/Per Shield")]
 public class DamagePerShield : DamageMod {
 
-    public override int Apply(int damage, UnitController caster, UnitController target) {
+    public override float Apply(float damage, UnitController caster, UnitController target) {
         return damage * caster.Shield;
     }
 }
